Moved method to service
